Working on debugging asynchronous actions - still full of bugs

+++ include/utility/driveMovement.hpp
@@ -15,7 +15,7 @@ class DriveMovement
 {
 private:
   int movementType;
-  int targetAngle;
+  double targetAngle;
   int speedDeadband;
   int maxSpeed;
   double kp;
