+++ include/utility/mathUtil.h
@@ -6,6 +6,6 @@ extern const double PI;
 extern double degreeToRadian(double degrees);
 extern double distance(double x1, double y1, double x2, double y2);
 extern bool isBetween(float number, float rangeLower, float rangeUpper);
-extern int sign(int num);
+extern double sign(double num);
 
 #endif
